working on ex01

+++ module 02/ex01/Fixed.hpp
@@ -6,7 +6,7 @@
 /*   By: hboumahd <hboumahd@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/11 14:31:59 by hboumahd          #+#    #+#             */
-/*   Updated: 2022/12/13 11:19:51 by hboumahd         ###   ########.fr       */
+/*   Updated: 2022/12/15 15:53:30 by hboumahd         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@
 class Fixed
 {
     private:
-        int fixedPoint;
+        float fixedPoint;
         static const int frctionlBits = 8;
     public:
         Fixed();
